Change the Software model texture format from 8-bit to 16-bit

+++ src/r_defs.h
@@ -749,7 +749,7 @@ typedef struct
 	INT16 width;
 	INT16 height;
 	UINT32 lumpnum;
-	UINT8 *data;
+	UINT16 *data;
 } rsp_spritetexture_t;
 #endif
